// react native/Mobile/WGFMobile/services/Model/ClipModel.tsx
export class ClipModel {
    playlistId: number = 0;
    playlistGuid: string = "";
    playlistName: string = "";
    clipId: number = 0;
    clipStart: number = 0;
    clipStop: number = 0;
    clipEnd: string = "";
    filmId: number = 0;
    clipServerName: string = "";
    clipClientName: string = "";
    filmGuid: string = "";
    pcId: number = 0;
    clipOrder: number = 0;
    clipLength: number = 0;
    clipTime: string = "";
    clipGuid: string = "";
    clipsdSize: number = 0;
    cliphdSize: number = 0;
    teamServer: string = "";
    teamDisk: string = "";
    teamGuid: string = "";
    angle1: number = 0;
    angleId1: number = 0;
    angle2: number = 0;
    angleId2: number = 0;
    angle3: number = 0;
    angleId3: number = 0;
    angle4: number = 0;
    angleId4: number = 0;
    angle5: number = 0;
    angle0: number = 0;
    dateCreated: string = "";
    dateUpdated: string = "";
    clipStatus: number = 0;
    custom1: string = "";
    custom2: string = "";
    custom3: string = "";
    miscGroup: string = "";
    miscTitle: string = "";
    miscCategory: string = "";
    sportPeriod: string = "";
    sportCategory: number = 0;
    sportGroup: string = "";
    sportTitle: string = "";
    sportResult: string = "";
    footballQtr: number = 0;
    footballOdk: string = "";
    footballPlayType: string = "";
    FootballOffPlay: string = "";
    footballOffForm: string = "";
    footballDown: number = 0;
    footballDistance: number = 0;
    footballGainLoss: number = 0;
    footballResult: string = "";
    footballYardLine: number = 0;
    volleyballServe: string = "";
    volleyballAce: string = "";
    volleyballDefense: string = "";
    volleyballSet: string = "";
    volleyballAssist: string = "";
    volleyballAttack: string = "";
    volleyballKill: string = "";
    volleyballBlock: string = "";
    volleyballError: string = "";
    volleyballPoint: number = 0;
    footballName: string = "";
    userId: number = 0;
    teamId: number = 0;
}